Module pallet_difttt::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use crate::weights::WeightInfo;

Modules

crypto

Structs

Pallet

The pallet implementing the on-chain logic.

RecipeDonePayload

Payload used by set recipe done to submit a transaction.

RecipeTimesPayload

Payload used by update recipe times to submit a transaction.

TradingPair

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

TokenSymbol

Constants

ACA

AUSD

BNC

DOT

KAR

KBTC

KEY_TYPE

KINT

KSM

KUSD

LCDOT

The first batch of lcDOT that expires at end of least 13

LDOT

LKSM

PHA

RENBTC

VSKSM

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

TokenInfo

Type Definitions

ActionOwner

AmountLimit

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextActionId

NextRecipeId

NextTrigerId

RecipeOwner

TrigerOwner

TxBlockLimit